fix(Dropdown): disabled toggle (#1571)
* fix: Dropdown disabled toggle issue

This fix addresses the issue where Dropdown will be passed a
`disabled` prop, yet it is still clickable. This is due to context
being passed the user-defined `toggle` prop directly, rather than
our defined `this.toggle` function that includes a check for
`disabled`.

Fixes #1542

* fix: incorrect test assertion

In the previous implementation, Dropdown wasn't properly firing
the toggle function within Dropdown, so the assertion was written
down wrong to compensate for that.