feat(UncontrolledDropdown): add onToggle callback (#1705)
Add `onToggle` to `UncontrolledDropdown`. The purpose of this new prop is to allow the user to be informed when the dropdown is toggled (opened/closed) as well as the new state of `isOpen` while still being uncontrolled.
when the dropdown is toggled, `onToggle` is called with the `event` and a `boolean` representing the new state (`true` mean the dropdown is being opened, `false` means it is being closed).